+++ aquadoggo/src/db/stores/document.rs
@@ -515,7 +515,7 @@ mod tests {
     use p2panda_rs::operation::traits::AsOperation;
     use p2panda_rs::operation::{Operation, OperationId};
     use p2panda_rs::storage_provider::traits::{DocumentStore, OperationStore};
-    use p2panda_rs::test_utils::constants::{self};
+    use p2panda_rs::test_utils::constants;
     use p2panda_rs::test_utils::fixtures::{
         operation, random_document_id, random_document_view_id, random_operation_id,
     };